got logprobs with vllm

+++ scripts/call_lm.py
@@ -1,19 +1,16 @@
+import os
 from argparse import ArgumentParser
+from glob import glob
 
 import pandas as pd
 from PIL import Image
 from pyprojroot import here
 
-from src.lm import get_logits
+# from src.lm import get_logits
+from src.lm_vllm import get_logits
 
 if __name__ == "__main__":
     parser = ArgumentParser()
-    parser.add_argument(
-        "--data_filepaths",
-        type=str,
-        default="data/trials_with_history.csv",
-        help="semicolon-separated list of data filepaths",
-    )
     parser.add_argument(
         "--model_name",
         type=str,
@@ -41,23 +38,30 @@ if __name__ == "__main__":
 
     args = parser.parse_args()
 
+    data_filepaths = glob(str(here("context_prep/*.csv")))
+    print("data filepaths:", data_filepaths)
+
     dfs = []
-    data_filepaths = args.data_filepaths.split(";")
     for filepath in data_filepaths:
         dfs.append(pd.read_csv(here(filepath)))
 
     grid_image = Image.open(here(args.grid_image_path))
 
-    result_dfs = get_logits(
-        dfs,
-        args.model_name,
-        grid_image,
-        n_trials=args.n_trials,
-        batch_size=args.batch_size,
-    )
-    
-    # Save results
-    for filepath, df in zip(data_filepaths, result_dfs):
-        output_path = filepath.replace(".csv", f"_{args.model_name.split('/')[-1]}_logprobs.csv")
+    for filepath, df in zip(data_filepaths, dfs):
+        output_path = filepath.replace(
+            ".csv", f"_{args.model_name.split('/')[-1]}_logprobs.csv"
+        )
+        if os.path.exists(output_path):
+            print(f"Skipping {filepath} as output file already exists.")
+            continue
+
+        print(f"Processing {filepath}...")
+        df_results = get_logits(
+            df,
+            args.model_name,
+            grid_image,
+            n_trials=args.n_trials,
+            batch_size=args.batch_size,
+        )
+
         df.to_csv(here(output_path), index=False)
-        print(f"Results saved to {output_path}")
